refactor(features): type stuff + easier imports
