+++ src/features/AntiPhishSystem.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import { ChannelSelectForm } from '@/components/forms/ChannelSelect';
-import { AntiPhishingFeature, UseFormRender } from '@/config/types';
+import type { AntiPhishingFeature } from '@/types/features';
+import type { UseFormRender } from '@/types/formTypes';
+import { ChannelSelectForm } from '@/components/forms';
 import { useForm } from 'react-hook-form';
 
 export const useAntiPhishFeature: UseFormRender<AntiPhishingFeature> = (data, onSubmit) => {
