+++ src/features/ConfessionsSystem.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import { ChannelSelectForm } from '@/components/forms/ChannelSelect';
-import { ConfessionsFeature, UseFormRender } from '@/config/types';
+import type { ConfessionsFeature } from '@/types/features';
+import type { UseFormRender } from '@/types/formTypes';
+import { ChannelSelectForm } from '@/components/forms';
 import { useForm } from 'react-hook-form';
 
 export const useConfessionSystem: UseFormRender<ConfessionsFeature> = (data, onSubmit) => {
