+++ src/features/TicketSystem.tsx
@@ -1,9 +1,8 @@
 'use client';
 
-import { ChannelSelectForm } from '@/components/forms/ChannelSelect';
-import { RoleSelectForm } from '@/components/forms/RoleSelect';
-import { TicketsFeature, UseFormRender } from '@/config/types';
-import Embed from '@/components/forms/Embed';
+import { ChannelSelectForm, RoleSelectForm, Embed } from '@/components/forms';
+import type { UseFormRender } from '@/types/formTypes';
+import type { TicketsFeature } from '@/types/features';
 import { useForm } from 'react-hook-form';
 import { Fragment } from 'react';
