+++ src/features/WelcomeSystem.tsx
@@ -1,9 +1,8 @@
 'use client';
 
-import { ChannelSelectForm } from '@/components/forms/ChannelSelect';
-import { TextAreaForm } from '@/components/forms/TextAreaForm';
-import { UseFormRender, WelcomeFeature } from '@/config/types';
-import Embed from '@/components/forms/Embed';
+import { ChannelSelectForm, TextAreaForm, Embed } from '@/components/forms';
+import type { WelcomeFeature } from '@/types/features';
+import type { UseFormRender } from '@/types/formTypes';
 import { useForm } from 'react-hook-form';
 import { Fragment } from 'react';
